LibWeb: Move WindowProxy from Bindings/ to HTML/
